refector(hdwallet): Rename 04-transaction/ethereum packages for better distinction
- Rename directory and package:
  * 04-transaction/ethereum -> 04-transaction/ethtx (package: ethtx)
