+++ hdwallet/04-transaction/ethtx/tx_test.go
@@ -1,4 +1,4 @@
-package ethereum
+package ethtx
 
 import (
 	"context"
